+++ ChangeLog.md
@@ -10,6 +10,7 @@
 * Pass through the timeout parameter from DicomClient.Send to the constructor of DesktopNetworkStream (#732)
 * Added appveyor.yml file for ci by this setting (#729)
 * Bug Fix : anonymized patient name is now encoded with same character set as the original Dicom.
+* add DicomUID.IsVolumeStorage.
 * Bug Fix : DICOM server may throw DicomDataException on association when non-standard transfer syntax was proposed (#749)
 * allow Query/Register/Unregister transfer syntax.